Split test/lib/db/listings.test.ts, drop its biome override
Removed test/lib/db/listings.test.ts (1188 lines) from biome.json's
noExcessiveLinesPerFile override list. Split it into five topic files
under 400 lines: listings-crud, listings-delete, listings-queries
(batch reads + name lookups + getStoredListingWithCount),
listings-transforms (write/read encryption transforms + duration
clamp + catalogVisibleSql), and listings-revenue.

The split let jscpd scan the file for the first time and surfaced four
duplicates:
- A "delete a listing, assert getListing returns null" test was a
  byte-for-byte duplicate of an earlier test in the same file (same
  setup, same assertion) — removed the redundant one.
- The "delete a listing, assert its attendees decrypt to none" pattern
  was duplicated against test/lib/db/attendees/decrypt-attendees.test.ts;
  extracted expectNoDecryptedAttendees into
  test-utils/db-helpers/attendees.ts and pointed both call sites at it.
- writeClosesAt/writeListingDate tests repeated the same
  import-decrypt-assert boilerplate; extracted a local decryptWritten
  helper in listings-transforms.test.ts.
- The revenue-breakdown reconciliation check (recognisedIncome vs the
  ledger projection, netBalance vs the account balance) was repeated
  across four tests; extracted a local expectReconciles helper in
  listings-revenue.test.ts.
